Library copying for native platforms (#472)
Adds library copying based on the platform that is being compiled.

Useful for copying things like steam sdks or other third party dlls.

For usage: have libraries in a folder that looks like this:
```
./         -- crate root
  lib/
    macos/
      libmylib.dylib
    linux/
      libmylib.so
    window/
      mylib.dll
      mylib.lib
```
and it'll be copied based on the platform to
```
./         -- application output directory
    app.exe
    assets/
    mylib.dll
    mylib.lib
```

Fixes https://github.com/TheBevyFlock/bevy_new_2d/issues/471